correção de tipo

+++ aula_220929/l02_2_5.cs
@@ -3,10 +3,10 @@ using System.Linq; // módulo permite o uso do Array.Selet() para conveter tipos
 public class Program {
   public static void Main() {
     Console.WriteLine("Digite o intervalo de tempo no formato “HH:MM:SS”");
-    int[] tempo = Console.ReadLine().Split(":").Select(int.Parse).ToArray();
+    long[] tempo = Console.ReadLine().Split(":").Select(long.Parse).ToArray();
     tempo[1] += tempo[0] * 60;
     tempo[2] += tempo[1] * 60;
-    int distancia = tempo[2] * 300000;
+    long distancia = tempo[2] * 300000;
     Console.WriteLine($"A luz percorreu {distancia} km nesse intervalo");
   }
 }
